chore(build): Kill ssh-agent in deploy script (#22)
The deploy script ran `ssh-agent` in the background, but because it was still running the Travis deploy script never finished and as such the Travis job never completed. This adds a step to kill `ssh-agent`.

The tag push logic (from `semantic-release`) will have to be run within the script so that travis will proceed to the next matrix. Getting rid of `deploy_key_decrypt.sh` in favor of just putting the commands out in the open in `.travis.yml`.